Add web control, birghtness
Add a HTTP function to switch the automatic aspect cycling on or
off, set the brightness and the current aspect. The home page
shows appropriate buttons that reflect the status at the time the
page was loaded.

Also add PWM output so we can turn down the brightness at night.